+++ app/src/main/java/app/learn/kotlin/di/module/builder/ActivityBuilder.kt
@@ -10,10 +10,8 @@ import app.learn.kotlin.feature.favourite.event.FavouriteEventFragment
 import app.learn.kotlin.feature.favourite.event.FavouriteEventModule
 import app.learn.kotlin.feature.favourite.team.FavouriteTeamFragment
 import app.learn.kotlin.feature.favourite.team.FavouriteTeamModule
-import app.learn.kotlin.feature.search.SearchActivity
 import app.learn.kotlin.feature.search.event.SearchEventFragment
 import app.learn.kotlin.feature.search.event.SearchEventModule
-import app.learn.kotlin.feature.search.SearchModule
 import app.learn.kotlin.feature.search.team.SearchTeamFragment
 import app.learn.kotlin.feature.search.team.SearchTeamModule
 import app.learn.kotlin.feature.team.detail.TeamDetailActivity
@@ -38,10 +36,6 @@ abstract class ActivityBuilder {
     @ActivityScope
     @ContributesAndroidInjector(modules = [MatchDetailModule::class])
     abstract fun provideMatchDetailActvity(): MatchDetailActivity
-
-    @ActivityScope
-    @ContributesAndroidInjector(modules = [SearchModule::class])
-    abstract fun provideSearchActivity(): SearchActivity
 
     @ActivityScope
     @ContributesAndroidInjector(modules = [SearchEventModule::class])
