--- app/src/main/java/app/learn/kotlin/feature/base/ContentView.kt
@@ -1,8 +1,0 @@
-package app.learn.kotlin.feature.base
-
-import android.view.View
-
-interface ContentView {
-
-    fun insertContentView(): View?
-}
