+++ app/src/main/java/app/learn/kotlin/feature/favourite/event/FavouriteEventFragment.kt
@@ -7,19 +7,17 @@ import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.ProgressBar
 import app.learn.kotlin.R
 import app.learn.kotlin.feature.HomeActivity
+import app.learn.kotlin.feature.adapter.EventAdapter
 import app.learn.kotlin.feature.base.BaseFragment
 import app.learn.kotlin.feature.event.detail.MatchDetailActivity
-import app.learn.kotlin.feature.adapter.EventAdapter
 import app.learn.kotlin.helper.invisible
 import app.learn.kotlin.helper.mapper
 import app.learn.kotlin.model.Constant
 import app.learn.kotlin.model.entity.EventEntity
 import app.learn.kotlin.model.vo.EventVo
 import com.airbnb.lottie.LottieAnimationView
-import kotlinx.android.synthetic.main.activity_match_detail.view.*
 import kotlinx.android.synthetic.main.fragment_favourite.view.*
 import kotlinx.android.synthetic.main.progress_bar.view.*
 import org.jetbrains.anko.startActivity
