+++ app/src/main/java/app/learn/kotlin/feature/search/SearchActivity.kt
@@ -2,10 +2,9 @@ package app.learn.kotlin.feature.search
 
 import android.os.Bundle
 import android.support.v4.app.Fragment
-import android.widget.ProgressBar
-import android.widget.SearchView
+import android.support.v7.app.AppCompatActivity
+import android.support.v7.widget.SearchView
 import app.learn.kotlin.R
-import app.learn.kotlin.feature.base.BaseActivity
 import app.learn.kotlin.feature.search.event.SearchEventContract
 import app.learn.kotlin.feature.search.event.SearchEventFragment
 import app.learn.kotlin.feature.search.team.SearchTeamContract
@@ -13,17 +12,10 @@ import app.learn.kotlin.feature.search.team.SearchTeamFragment
 import app.learn.kotlin.model.Constant
 import app.learn.kotlin.model.Constant.FRAGMENT_EVENT
 import app.learn.kotlin.model.Constant.FRAGMENT_TEAM
-import app.learn.kotlin.model.response.Event
-import app.learn.kotlin.model.response.ListResponse
-import app.learn.kotlin.model.response.Team
 import com.airbnb.lottie.LottieAnimationView
 import kotlinx.android.synthetic.main.fragment_search.*
-import javax.inject.Inject
 
-class SearchActivity : BaseActivity<SearchContract.Presenter> (), SearchContract.View {
-
-    @Inject
-    internal lateinit var presenter: SearchContract.Presenter
+class SearchActivity : AppCompatActivity() {
 
     private lateinit var searchEvent: SearchEventContract.View
     private lateinit var searchTeam: SearchTeamContract.View
@@ -31,7 +23,8 @@ class SearchActivity : BaseActivity<SearchContract.Presenter> (), SearchContract
     private lateinit var progressBar: LottieAnimationView
     private lateinit var flag: String
 
-    override fun onInitView() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
         setContentView(R.layout.fragment_search)
         btn_close.setOnClickListener {
             onBackPressed()
@@ -41,7 +34,7 @@ class SearchActivity : BaseActivity<SearchContract.Presenter> (), SearchContract
         search.setIconifiedByDefault(false)
         search.requestFocus()
 
-        search.setOnQueryTextListener(object : SearchView.OnQueryTextListener, android.support.v7.widget.SearchView.OnQueryTextListener {
+        search.setOnQueryTextListener(object : SearchView.OnQueryTextListener {
             override fun onQueryTextSubmit(query: String?): Boolean {
                 search(query.orEmpty())
                 return false
@@ -78,14 +71,6 @@ class SearchActivity : BaseActivity<SearchContract.Presenter> (), SearchContract
         }
     }
 
-    override fun addTeamsFound(teams: ListResponse<Team>) {
-
-    }
-
-    override fun addEventsFound(events: ListResponse<Event>) {
-
-    }
-
     private fun <T : Fragment> showFragment(savedInstanceState: Bundle?, fragment: T) {
         if (savedInstanceState == null) {
             supportFragmentManager
@@ -94,8 +79,4 @@ class SearchActivity : BaseActivity<SearchContract.Presenter> (), SearchContract
                     .commit()
         }
     }
-
-    override fun getPresenter(): SearchContract.Presenter? = presenter
-
-    override fun getProgressBar(): LottieAnimationView?  = progressBar
 }
