+++ app/src/main/java/app/learn/kotlin/feature/search/event/SearchEventContract.kt
@@ -4,6 +4,7 @@ import app.learn.kotlin.feature.base.BasePresenter
 import app.learn.kotlin.feature.base.BaseView
 import app.learn.kotlin.model.response.Event
 import app.learn.kotlin.model.response.ListResponse
+
 class SearchEventContract {
 
     interface  View : BaseView {
