+++ app/src/main/java/app/learn/kotlin/feature/search/team/SearchTeamFragment.kt
@@ -6,11 +6,10 @@ import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.ProgressBar
 import app.learn.kotlin.R
+import app.learn.kotlin.feature.adapter.TeamAdapter
 import app.learn.kotlin.feature.base.BaseFragment
 import app.learn.kotlin.feature.team.detail.TeamDetailActivity
-import app.learn.kotlin.feature.adapter.TeamAdapter
 import app.learn.kotlin.helper.invisible
 import app.learn.kotlin.helper.mapper
 import app.learn.kotlin.model.Constant
@@ -19,7 +18,6 @@ import app.learn.kotlin.model.response.Team
 import app.learn.kotlin.model.vo.TeamVo
 import com.airbnb.lottie.LottieAnimationView
 import kotlinx.android.synthetic.main.base_recycle_view.view.*
-import kotlinx.android.synthetic.main.fragment_match.view.*
 import kotlinx.android.synthetic.main.progress_bar.view.*
 import org.jetbrains.anko.startActivity
 import org.jetbrains.anko.support.v4.ctx
@@ -39,7 +37,7 @@ class SearchTeamFragment : BaseFragment<SearchTeamContract.Presenter>(), SearchT
     override fun getPresenter(): SearchTeamContract.Presenter? = presenter
 
     override fun onInitView(inflater: LayoutInflater?, container: ViewGroup?, savedInstanceState: Bundle?): View {
-        val view = layoutInflater.inflate(R.layout.fragment_search_team, container, false);
+        val view = layoutInflater.inflate(R.layout.fragment_search_team, container, false)
         progressBar = view.base_progress_bar_id
         progressBar.invisible()
         recycleView = view.base_recycle_view_id
