+++ app/src/main/java/app/learn/kotlin/feature/search/team/SearchTeamModule.kt
@@ -1,8 +1,6 @@
 package app.learn.kotlin.feature.search.team
 
 import app.learn.kotlin.di.scope.ActivityScope
-import app.learn.kotlin.feature.search.event.SearchEventContract
-import app.learn.kotlin.feature.search.event.SearchEventPresenterImpl
 import dagger.Binds
 import dagger.Module
