+++ app/src/main/java/app/learn/kotlin/feature/team/player/PlayerDetailActivity.kt
@@ -5,8 +5,6 @@ import android.support.design.widget.AppBarLayout
 import android.support.design.widget.CollapsingToolbarLayout
 import android.support.v7.app.AppCompatActivity
 import android.view.MenuItem
-import app.learn.kotlin.R
-import app.learn.kotlin.R.id.view_pager_player
 import app.learn.kotlin.R.layout.activity_player_detail
 import app.learn.kotlin.feature.base.BasePagerAdapter
 import app.learn.kotlin.helper.loadImageUrl
