--- app/src/main/res/anim/anim_rotate.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<rotate
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:fromDegrees="0"
-    android:interpolator="@android:anim/linear_interpolator"
-    android:toDegrees="360"
-    android:pivotX="50%"
-    android:pivotY="50%"
-    android:duration="500"
-    android:startOffset="0"
-    android:repeatCount="infinite" />
