--- app/src/main/res/drawable/selected_favorite_state.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item
-        android:drawable="@drawable/ic_unselected_favorite"
-        android:state_selected="false" />
-    <item
-        android:drawable="@drawable/ic_selected_favorite"
-        android:state_selected="true" />
-</selector>
