+++ app/src/main/res/layout/fragment_player_overview.xml
@@ -78,7 +78,7 @@
 
         <android.support.v4.widget.NestedScrollView
             android:layout_width="match_parent"
-            android:layout_height="match_parent">
+            android:layout_height="wrap_content">
             <TextView
                 android:id="@+id/tv_player_description"
                 android:layout_width="match_parent"
