+++ app/src/main/res/layout/fragment_team_overview.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent">
     <android.support.v4.widget.NestedScrollView
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="wrap_content">
         <TextView
             android:id="@+id/tv_overview"
             android:layout_width="match_parent"
