+++ app/src/main/res/values/colors.xml
@@ -7,6 +7,5 @@
     <color name="gray">#e6e6e6</color>
     <color name="darkGray">#606060</color>
     <color name="colorAccent">#D81B60</color>
-    <color name="red">#f70d34</color>
     <color name="black">#000</color>
 </resources>
