+++ app/src/main/res/values/styles.xml
@@ -8,7 +8,4 @@
         <item name="colorAccent">@color/colorAccent</item>
     </style>
 
-    <style name="HeaderBar">
-        <item name="android:background">#303F9F</item>
-    </style>
 </resources>
